backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-multipart==0.0.20
python-dotenv==1.0.1
google-genai==1.4.0
openpyxl==3.1.5
python-docx==1.1.2
pypdf==5.1.0
extract-msg==0.52.0
aiofiles==24.1.0
pydantic==2.10.0
pydantic-settings==2.6.0
beautifulsoup4==4.12.3
